+++ Content.Shared/VendingMachines/VendingMachineInventoryPrototype.cs
@@ -1,23 +1,21 @@
 using Robust.Shared.Prototypes;
-using Robust.Shared.Serialization;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype.Dictionary;
 
-namespace Content.Shared.VendingMachines
+namespace Content.Shared.VendingMachines;
+
+[Prototype]
+public sealed partial class VendingMachineInventoryPrototype : IPrototype
 {
-    [Prototype]
-    public sealed partial class VendingMachineInventoryPrototype : IPrototype
-    {
-        [ViewVariables]
-        [IdDataField]
-        public string ID { get; private set; } = default!;
+    /// <inheritdoc/>
+    [IdDataField]
+    public string ID { get; private set; } = default!;
 
-        [DataField("startingInventory", customTypeSerializer:typeof(PrototypeIdDictionarySerializer<uint, EntityPrototype>))]
-        public Dictionary<string, uint> StartingInventory { get; private set; } = new();
+    [DataField]
+    public Dictionary<ProtoId<EntityPrototype>, uint> StartingInventory { get; private set; } = [];
 
-        [DataField("emaggedInventory", customTypeSerializer:typeof(PrototypeIdDictionarySerializer<uint, EntityPrototype>))]
-        public Dictionary<string, uint>? EmaggedInventory { get; private set; }
+    [DataField]
+    public Dictionary<ProtoId<EntityPrototype>, uint>? EmaggedInventory { get; private set; }
 
-        [DataField("contrabandInventory", customTypeSerializer:typeof(PrototypeIdDictionarySerializer<uint, EntityPrototype>))]
-        public Dictionary<string, uint>? ContrabandInventory { get; private set; }
-    }
+    [DataField]
+    public Dictionary<ProtoId<EntityPrototype>, uint>? ContrabandInventory { get; private set; }
 }
